CORE-8977 Move two more Utils files
