+++ applications/workers/workers-smoketest/src/smokeTest/kotlin/net/corda/applications/workers/smoketest/ConfigTests.kt
@@ -1,6 +1,11 @@
 package net.corda.applications.workers.smoketest
 
+import net.corda.e2etest.utilities.configWithDefaultsNode
+import net.corda.e2etest.utilities.getConfig
+import net.corda.e2etest.utilities.sourceConfigNode
 import net.corda.e2etest.utilities.toJsonString
+import net.corda.e2etest.utilities.updateConfig
+import net.corda.e2etest.utilities.waitForConfigurationChange
 import net.corda.schema.configuration.ConfigKeys
 import net.corda.schema.configuration.ConfigKeys.RECONCILIATION_CONFIG
 import net.corda.schema.configuration.ReconciliationConfig
