+++ testing/e2e-test-utilities/src/main/kotlin/net/corda/e2etest/utilities/RbacTestUtils.kt
@@ -1,13 +1,8 @@
-package net.corda.applications.workers.smoketest
+package net.corda.e2etest.utilities
 
 import java.time.Instant
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
-import net.corda.e2etest.utilities.CLUSTER_URI
-import net.corda.e2etest.utilities.PASSWORD
-import net.corda.e2etest.utilities.USERNAME
-import net.corda.e2etest.utilities.assertWithRetry
-import net.corda.e2etest.utilities.cluster
 
 object RbacTestUtils {
